requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
spacy>=3.7.0
argostranslate>=1.9.0
